Cater for nested aliassed types
