+++ packages/types/src/interfaces/proxy/definitions.ts
@@ -22,8 +22,6 @@ export default {
       real: 'AccountId',
       callHash: 'Hash',
       height: 'BlockNumber'
-    },
-    // FIXME We don't want this here
-    Announcement: 'ProxyAnnouncement'
+    }
   }
 } as Definitions;
